Port block world into Vite demo
